# docs/swagger.yaml
definitions:
...
  dto.ChannelListItemOutput:
    properties:
      channel_id:
        description: 通道ID
        type: string
      create_at:
        description: 创建时间
        type: string
      detail:
        description: 通道描述
        type: string
      id:
        description: id
        type: integer
      state:
        description: 状态
        type: string
      total_node:
        description: 节点数
        type: integer
    type: object
  dto.ChannelListOutput:
    properties:
      list:
        description: 列表
        items:
          $ref: '#/definitions/dto.ChannelListItemOutput'
        type: array
      total:
        description: 总数
        type: integer
    type: object
  dto.ContentListItemOutput:
    properties:
      content_name:
        description: 合约名称
        type: string
      detail:
        description: detail
        type: string
      id:
        description: id
        type: integer
      service_name:
        description: ServiceName
        type: string
      service_type:
        description: ServiceType
        type: string
      tap:
        description: tap
        type: string
    type: object
  dto.ContentListOutput:
    properties:
      list:
        description: 列表
        items:
          $ref: '#/definitions/dto.ContentListItemOutput'
        type: array
      total:
        description: 总数
        type: integer
    type: object
...
  dto.DhtOutput:
    properties:
      dhtId:
        type: string
      humitidy:
        type: integer
      id:
        type: integer
      temperature:
        type: integer
      time:
        type: string
    type: object
  dto.DhtSetting:
    properties:
      dhtId:
        type: string
      nowSample:
        type: integer
    type: object
...
  dto.TokensInput:
    properties:
      grant_type:
        description: 授权类型
        example: client_credentials
        type: string
      scope:
        description: 权限范围
        example: read_write
        type: string
    required:
    - grant_type
    - scope
    type: object
  dto.TokensOutput:
    properties:
      access_token:
        description: access_token
        type: string
      expires_in:
        description: expires_in
        type: integer
      scope:
        description: scope
        type: string
      token_type:
        description: token_type
        type: string
    type: object
  middleware.Response:
    properties:
      data: {}
      errmsg:
        type: string
      errno:
        type: integer
      stack: {}
      trace_id: {}
    type: object
info:
  contact: {}
paths:
...
  /channel/channel_count:
    get:
      consumes:
      - application/json
      description: 通道数
      operationId: /channel/channel_count
      parameters:
      - description: 通道ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 通道数
      tags:
      - 通道管理
...
  /channel/channel_list:
    get:
      consumes:
      - application/json
      description: 通道列表
      operationId: /channel/channel_list
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页个数
        in: query
        name: page_size
        required: true
        type: integer
      - description: 当前页数
        in: query
        name: page_no
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ChannelListOutput'
              type: object
      summary: 通道列表
      tags:
      - 通道管理
...
  /channel/content/dhtdashboard/dhtdata_get:
    get:
      consumes:
      - application/json
      description: 取得所有dhtid输出温度数据形式
      operationId: /channel/content/dhtdashboard/dhtdata_get
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页个数
        in: query
        name: page_size
        type: integer
      - description: 当前页数
        in: query
        name: page_no
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.DhtOutput'
              type: object
      summary: 取得所有dhtid输出温度数据形式
      tags:
      - dht接口
  /channel/content/dhtdashboard/dhtsetting:
    post:
      consumes:
      - application/json
      description: Dhtsetting
      operationId: /channel/content/dhtdashboard/dhtsetting
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.DhtSetting'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Dhtsetting
      tags:
      - dht接口
...
  /channel/content_list:
    get:
      consumes:
      - application/json
      description: 合约列表
      operationId: /channel/content_list
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页个数
        in: query
        name: page_size
        required: true
        type: integer
      - description: 当前页数
        in: query
        name: page_no
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ContentListOutput'
              type: object
      summary: 合约列表
      tags:
      - 通道管理
...
  /oauth/tokens:
    post:
      consumes:
      - application/json
      description: 获取TOKEN
      operationId: /oauth/tokens
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TokensInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TokensOutput'
              type: object
      summary: 获取TOKEN
      tags:
      - OAUTH
swagger: "2.0"